Development

+++ api/src/wsa_sweep_device.c
@@ -19,6 +19,7 @@
 #include "wsa_lib.h"
 #include "wsa_dsp.h"
 #include "wsa_debug.h"
+#include "wsa_lib.h"
 #ifndef _TIMES_H
 #define _TIMES_H
 
@@ -751,10 +752,10 @@ static int wsa_plan_sweep(struct wsa_sweep_device *sweep_device, struct wsa_powe
 	// do we need a cleanup entry?
 	if ((fcstop + half_usable_bw) < pscfg->fstop) {
 		// how much is left over? (it should be less than usable_bw)
-		tmpfreq = (float) (pscfg->fstop - (fcstop + half_usable_bw));
+		tmpfreq = (float)(pscfg->fstop - (fcstop + half_usable_bw));
 
 		// make a sweep entry for fcstop + half of that
-		tmpfreq = pscfg->fstop - (tmpfreq / 2);
+		tmpfreq = (float)(pscfg->fstop - (tmpfreq / 2));
 
 		// which must be a multiple of freq resolution
 		tmpfreq = (tmpfreq / prop->tuning_resolution) * prop->tuning_resolution;
@@ -766,7 +767,6 @@ static int wsa_plan_sweep(struct wsa_sweep_device *sweep_device, struct wsa_powe
 		new_entry_freq = (uint64_t) tmpfreq;
 		pscfg->sweep_plan->next_entry = wsa_sweep_plan_entry_new(new_entry_freq, new_entry_freq, fstep, points, 1, dd_mode);
 	}
-
 	
 	// how many steps on in this plan? loop through the list and count 'em
 
@@ -781,8 +781,7 @@ static int wsa_plan_sweep(struct wsa_sweep_device *sweep_device, struct wsa_powe
 			pscfg->packet_total += plan->ppb;
 		}
 		else {
-			pscfg->packet_total += (uint64_t) ((((plan->fcstop - plan->fcstart) /  plan->fstep) + 1) * (uint32_t) (plan->ppb));
-
+			pscfg->packet_total += (uint32_t)( (((plan->fcstop - plan->fcstart) /  plan->fstep) + 1) * plan->ppb);
 		}
 	}
